fix: resolve all compile errors after upstream merge
- Remove McpServerTransportConfig::Zap (variant removed upstream)
- Fix ratatui 0.29 Color type conversions (ratatui::Color vs crossterm::Color)
- Fix Cell<u16> scroll_offset API in layout_scroll.rs
- Add missing modules (intro_art_variants, request_user_input_view, external_editor)
- Fix zen field type (bool → Option<ZenThemeConfig>)
- Add GutterMode::next() method
- Remove deleted SlashCommand::Rename, add Provider/Share stubs
- Fix schedule_cli_prompt signature (3 args, store overrides on controller)
- Fix ModelInfoUpgrade field access and default_reasoning_level Option handling
- Align cloud-tasks deps with workspace (ratatui 0.29, throbber-widgets-tui 0.8)
- Add missing auth constants and version functions
- Replace wire_api with dynamic_tools in ConfigOverrides
- Add end_turn/phase fields to ResponseItem::Message
- Fix format_with_separators to accept i64

+++ hanzo-dev/core/src/image_comparison.rs
@@ -1,9 +1,6 @@
 // Use img_hash's re-exported image module for compatibility
-use img_hash::HashAlg;
-use img_hash::HasherConfig;
-use img_hash::ImageHash;
-use img_hash::image::DynamicImage;
-use img_hash::image::io::Reader as ImageReader;
+use img_hash::image::{DynamicImage, io::Reader as ImageReader};
+use img_hash::{HashAlg, HasherConfig, ImageHash};
 use std::path::Path;
 
 fn phash_256(img: &DynamicImage) -> ImageHash<[u8; 32]> {
@@ -44,13 +41,13 @@ pub fn are_hashes_similar(phash1: &[u8], dhash1: &[u8], phash2: &[u8], dhash2: &
     let phash_dist = phash1
         .iter()
         .zip(phash2.iter())
-        .map(|(a, b)| (a ^ b).count_ones())
+        .map(|(a, b)| (a ^ b).count_ones() as u32)
         .sum::<u32>();
 
     let dhash_dist = dhash1
         .iter()
         .zip(dhash2.iter())
-        .map(|(a, b)| (a ^ b).count_ones())
+        .map(|(a, b)| (a ^ b).count_ones() as u32)
         .sum::<u32>();
 
     // 256 bits → ~5% tolerance (≈13 bits)
